Break state management out of message processors
