+++ src/main/java/kitschinfernologger/FileLoggerMessageProcessor.java
@@ -14,18 +14,16 @@ import static net.runelite.client.RuneLite.RUNELITE_DIR;
 @Slf4j
 public class FileLoggerMessageProcessor extends BaseMessageProcessor {
     @Override
-    protected void HandleCompletionMessage(ChatMessage message) {
-        super.HandleCompletionMessage(message);
-        writeSplitsToFile();
+    public void onCompletionMessage(ChatMessage message, InfernoState state) {
+        writeSplitsToFile(state);
     }
 
     @Override
-    protected void HandleDefeatedMessage(ChatMessage message) {
-        super.HandleDefeatedMessage(message);
-        writeSplitsToFile();
+    public void onDefeatedMessage(ChatMessage message, InfernoState state) {
+        writeSplitsToFile(state);
     }
 
-    private void writeSplitsToFile() {
+    private void writeSplitsToFile(InfernoState state) {
         if (!config.getShouldWriteToFile()) {
             return;
         }
@@ -33,7 +31,7 @@ public class FileLoggerMessageProcessor extends BaseMessageProcessor {
         File dir = new File(RUNELITE_DIR, "InfernoTimerLogs/" + client.getLocalPlayer().getName());
         dir.mkdirs();
 
-        String fileName = getFileName();
+        String fileName = getFileName(state);
         try (FileWriter fw = new FileWriter(new File(dir, fileName)))
         {
             fw.write(state.getSplitsCsv());
@@ -44,7 +42,7 @@ public class FileLoggerMessageProcessor extends BaseMessageProcessor {
         }
     }
 
-    private String getFileName() {
+    private String getFileName(InfernoState state) {
         if (state.getKillCount() == 0) {
             DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy-MM-dd HH;mm");
             LocalDateTime now = LocalDateTime.now();
